Move Telegram lexicon to presentation layer
