+++ presentation/telegram/lexicon.py
@@ -1,3 +1,6 @@
+"""Telegram presentation lexicon helper."""
+from __future__ import annotations
+
 from typing import Dict
 
 _LEXICON: Dict[str, Dict[str, str]] = {
@@ -17,6 +20,7 @@ _LEXICON: Dict[str, Dict[str, str]] = {
 
 
 def lexeme(key: str, lang: str = "en") -> str:
+    """Return a localized lexeme for Telegram presentation."""
     lang_norm = (lang or "en").split("-")[0].lower()
     d = _LEXICON.get(key, {})
     return d.get(lang_norm) or d.get("en") or ""
